Separated - WORKING
Separated into smaller chunks, demonstrating how it would likely be used in a workshop.
